comment question

+++ src/main/java/com/life/hz/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     public String editQuestion( @RequestParam(value = "title",required=false)String title,
                                 @RequestParam(value = "desciption",required=false)String desciption,
                                 @RequestParam(value = "tag",required=false)String tag,
-                                @PathVariable(name = "id",required=false)Integer id,
+                                @PathVariable(name = "id",required=false)Long id,
                                 Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("title",questionDTO.getTitle());
@@ -44,7 +44,7 @@ public class PublishController {
             @RequestParam(value = "title",required=false)String title,
             @RequestParam(value = "desciption",required=false)String desciption,
             @RequestParam(value = "tag",required=false)String tag,
-            @RequestParam(value = "id",required = false)Integer id,
+            @RequestParam(value = "id",required = false)Long id,
             HttpServletRequest request,
             Model model){
